E2E tests - added register test, fixed some sonarcloud issues

+++ server/api/user.js
@@ -94,14 +94,20 @@ router.get('/logout', auth, (req, res) => {
     .json({ message: 'Successfully logged out' })
 })
 
-router.post('/delete', async (req, res) => {
+// delete user
+router.delete('/delete/:id', auth, async (req, res) => {
   try {
-    const { email } = req.body
-    // Verify email
-    const user = await User.deleteOne({ email: email })
-    return res.status(200).json({
-      message: 'User deleted successfully',
-    })
+    const user = await User.findOne({ _id: req.userId })
+    const email = String(req.params.id)
+    // check admin role
+    if (user.role == 'admin') {
+      // verify email
+      const user_email = await User.deleteOne({ email: email })
+      // delete user
+      return res.status(200).json({
+        message: 'User deleted successfully',
+      })
+    }
   } catch (err) {
     res.status(500).send('Server error')
   }
